Side test ab minus iefix (#2131)
Since the final commit in #2130 to fix IE stuff is what broke Travis (and it seems the current "IE is running AB tests" is in production, I want to land this branch, which is #2130 without its most recent commit, plus reduction of the AB test implemented here to 0% while the other test finishes running. This way localizers can get their hands on this prior to flipping the switch next week. 

cc @Pomax